Put getCacheControlDirectivesFromConfig into its own file
