+++ plugins/core/base/index.js
@@ -1,4 +1,5 @@
 const getToolResponse = require("./methods/getToolResponse.js").getToolResponse;
+const getCacheControlDirectivesFromConfig = require("./methods/getCacheControlDirectivesFromConfig.js").getCacheControlDirectivesFromConfig;
 
 module.exports = {
   name: "q-base",
@@ -6,38 +7,7 @@ module.exports = {
   register: async function (server, options) {
     await server.register([require("vision"), require("inert")]);
     server.method("getToolResponse", getToolResponse);
-    server.method("getCacheControlDirectivesFromConfig", async function (
-      cacheControlConfig
-    ) {
-      // return early if no config given, default is 'no-cache'
-      if (!cacheControlConfig) {
-        return ["no-cache"];
-      }
-
-      const cacheControlDirectives = [];
-
-      if (cacheControlConfig.public) {
-        cacheControlDirectives.push("public");
-      }
-      if (cacheControlConfig.maxAge) {
-        cacheControlDirectives.push(`max-age=${cacheControlConfig.maxAge}`);
-      }
-      if (cacheControlConfig.sMaxAge) {
-        cacheControlDirectives.push(`s-maxage=${cacheControlConfig.sMaxAge}`);
-      }
-      if (cacheControlConfig.staleWhileRevalidate) {
-        cacheControlDirectives.push(
-          `stale-while-revalidate=${cacheControlConfig.staleWhileRevalidate}`
-        );
-      }
-      if (cacheControlConfig.staleIfError) {
-        cacheControlDirectives.push(
-          `stale-if-error=${cacheControlConfig.staleIfError}`
-        );
-      }
-
-      return cacheControlDirectives;
-    });
+    server.method("getCacheControlDirectivesFromConfig", getCacheControlDirectivesFromConfig);
 
     server.event("item.new");
     server.event("item.update");
